refactor: move all packet classes to separate files
