+++ afp-ui/Packet/Data/MacroRecordPacketData.cs
@@ -1,0 +1,7 @@
+namespace AFP.Packet;
+
+public class MacroRecordPacketData : IPacketData
+{
+	public required bool Record { get; set; }
+	public required ushort Slot { get; set; }
+}
