+++ afp-ui/Packet/Data/MacroStatePacketData.cs
@@ -1,0 +1,6 @@
+namespace AFP.Packet;
+
+public class MacroStatePacketData : IPacketData
+{
+	public required bool[] Macros { get; set; }
+}
