+++ afp-ui/Packet/Data/ScopeActionPacketData.cs
@@ -1,0 +1,7 @@
+namespace AFP.Packet;
+
+public class ScopeActionPacketData : IPacketData
+{
+	public required string Action { get; set; }
+	public required string Scope { get; set; }
+}
