+++ afp-ui/Packet/Data/ScopeInfoPacketData.cs
@@ -1,0 +1,6 @@
+namespace AFP.Packet;
+
+public class ScopeInfoPacketData : IPacketData
+{
+	public required ushort ChannelCount { get; set; }
+}
