+++ afp-ui/Packet/Data/ScopeListPacketData.cs
@@ -1,0 +1,6 @@
+namespace AFP.Packet;
+
+public class ScopeListPacketData : IPacketData
+{
+	public required string[] Scopes { get; set; }
+}
