+++ afp-ui/Packet/ScopeStatus.cs
@@ -1,0 +1,8 @@
+namespace AFP.Packet;
+
+public enum ScopeStatus
+{
+	Connected,
+	Disconnected,
+	Connecting,
+}
